% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.PICPWf.R
\name{calc.PIC}
\alias{calc.PIC}
\title{Calculate PIC}
\usage{
calc.PIC(x, y, z, nnmax = 10000, nvarmax = 100)
}
\arguments{
\item{x}{A vector of response.}

\item{y}{A vector of new predictor.}

\item{z}{A matrix of pre-existing predictors that could be NULL if no prior predictors exist.}

\item{nnmax}{The maximum number of sample size.}

\item{nvarmax}{The maximum number of potential predictors.}
}
\value{
A list of 2 elements: the partial mutual information (pmi), and partial informational correlation (pic).
}
\description{
Calculate PIC
}
\references{
Sharma, A., Mehrotra, R., 2014. An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1): 650-660.
}
