% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.PICPWf.R
\name{calc.PW}
\alias{calc.PW}
\title{Calculate Partial Weight}
\usage{
calc.PW(x, py, cpy, cpyPIC)
}
\arguments{
\item{x}{A vector of response.}

\item{py}{A matrix containing possible predictors of x.}

\item{cpy}{The column numbers of the meaningful predictors (cpy).}

\item{cpyPIC}{Partial informational correlation (cpyPIC).}
}
\value{
A vector of partial weights(pw) of the same length of z.
}
\description{
Calculate Partial Weight
}
\references{
Sharma, A., Mehrotra, R., 2014. An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1): 650-660.
}
