% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{data1}
\alias{data1}
\title{Sample data : AR9 model:  x(i)=0.3*x(i-1)-0.6*x(i-4)-0.5*x(i-9)+eps}
\usage{
data(data1)
}
\description{
A dataset containing 500 rows (data length) and 16 columns. The first column is response data and the rest columns are possible predictors.
}
\keyword{datasets}
