% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{data2}
\alias{data2}
\title{Sample data : AR4 model:  x(i)=0.6*x(i-1)-0.4*x(i-4)+eps}
\usage{
data(data2)
}
\description{
A dataset containing 500 rows (data length) and 16 columns. The first column is response data and the rest columns are possible predictors.
}
\keyword{datasets}
