% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{data3}
\alias{data3}
\title{Sample data : AR1 model:  x(i)=0.9*x(i-1)+0.866*eps}
\usage{
data(data3)
}
\description{
A dataset containing 500 rows (data length) and 16 columns. The first column is response data and the rest columns are possible predictors.
}
\keyword{datasets}
