% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.PICPW.R
\name{pic.calc}
\alias{pic.calc}
\title{Calculate PIC}
\usage{
pic.calc(X, Y, Z)
}
\arguments{
\item{X}{A vector of response.}

\item{Y}{A matrix of new predictors.}

\item{Z}{A matrix of pre-existing predictors that could be NULL if no prior predictors exist.}
}
\value{
A list of 2 elements: the partial mutual information (pmi), and partial informational correlation (pic).
}
\description{
Calculate PIC
}
\references{
Sharma, A., Mehrotra, R., 2014. An information theoretic alternative to model a natural system using observational information alone. Water Resources Research, 50(1): 650-660.
}
