% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{Sydney}
\alias{Sydney}
\title{Sample data: Data over Sydney region}
\usage{
data(Sydney)
}
\description{
A dataset containing Rainfall (15 stations), NCEP and CSIRO (7 atmospheric variables).
}
\keyword{datasets}
